test(actions): split registry/validator/pipeline tests into per-feature submodules (#494)
Completes the #494 split: the three remaining offenders (registry/tests.rs 648 LOC, pipeline/tests.rs 598 LOC, validator/tests.rs 513 LOC) are each converted to a directory with one file per feature/section + a fixtures.rs for shared setup.

registry/tests/ — 7 split files, largest 174 LOC (happy_path), shared fixtures.rs with ActionMeta builders. pipeline/tests/ — 9 split files, largest 192 LOC (custom), shared fixtures.rs. validator/tests/ — 5 split files, largest 168 LOC (advanced), shared fixtures.rs.

Behavioural parity: 238 tests pass (same as the baseline before split). cargo fmt + cargo clippy -p dcc-mcp-actions --all-targets -- -D warnings clean.

All five files originally listed in the issue (loader, catalog, registry, pipeline, validator) are now under 300 LOC per file. Closes #494.
