+++ crates/dcc-mcp-actions/src/registry/tests/mod.rs
@@ -1,0 +1,8 @@
+use super::*;
+
+mod batch_unregister;
+mod concurrency;
+pub(super) mod fixtures;
+mod happy_path;
+mod search;
+mod serialization;
