--- crates/dcc-mcp-actions/src/validator/tests.rs
@@ -1,513 +1,0 @@
-use super::*;
-use serde_json::json;
-
-use crate::registry::ActionMeta;
-
-fn make_meta_with_schema(schema: Value) -> ActionMeta {
-    ActionMeta {
-        name: "test_action".into(),
-        dcc: "maya".into(),
-        input_schema: schema,
-        ..Default::default()
-    }
-}
-
-// ── type checks ───────────────────────────────────────────────────────────
-
-#[test]
-fn test_valid_object_passes() {
-    let meta = make_meta_with_schema(json!({ "type": "object" }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!({})).is_valid());
-    assert!(v.validate_input(&json!({"x": 1})).is_valid());
-}
-
-#[test]
-fn test_wrong_type_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "object" }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!("hello")).is_valid());
-    assert!(!v.validate_input(&json!(42)).is_valid());
-    assert!(!v.validate_input(&json!(null)).is_valid());
-}
-
-#[test]
-fn test_number_type_accepts_int_and_float() {
-    let meta = make_meta_with_schema(json!({ "type": "number" }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(1.5)).is_valid());
-    assert!(v.validate_input(&json!(42)).is_valid());
-}
-
-#[test]
-fn test_integer_type_rejects_float() {
-    let meta = make_meta_with_schema(json!({ "type": "integer" }));
-    let v = ActionValidator::new(&meta);
-    // serde_json represents 1.5 as f64, not integer
-    assert!(v.validate_input(&json!(42)).is_valid());
-    // 1.5 will be f64
-    assert!(!v.validate_input(&json!(1.5)).is_valid());
-}
-
-#[test]
-fn test_boolean_type() {
-    let meta = make_meta_with_schema(json!({ "type": "boolean" }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(true)).is_valid());
-    assert!(!v.validate_input(&json!(1)).is_valid());
-}
-
-#[test]
-fn test_null_type() {
-    let meta = make_meta_with_schema(json!({ "type": "null" }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(null)).is_valid());
-    assert!(!v.validate_input(&json!(0)).is_valid());
-}
-
-#[test]
-fn test_union_type() {
-    let meta = make_meta_with_schema(json!({ "type": ["string", "null"] }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!("hello")).is_valid());
-    assert!(v.validate_input(&json!(null)).is_valid());
-    assert!(!v.validate_input(&json!(1)).is_valid());
-}
-
-// ── required ──────────────────────────────────────────────────────────────
-
-#[test]
-fn test_required_field_present() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "required": ["name"],
-        "properties": { "name": { "type": "string" } }
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!({"name": "sphere"})).is_valid());
-}
-
-#[test]
-fn test_required_field_missing() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "required": ["name"],
-        "properties": { "name": { "type": "string" } }
-    }));
-    let v = ActionValidator::new(&meta);
-    let result = v.validate_input(&json!({}));
-    assert!(!result.is_valid());
-    assert!(
-        result.errors[0]
-            .message
-            .contains("missing required field 'name'")
-    );
-}
-
-#[test]
-fn test_multiple_required_fields_missing() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "required": ["x", "y", "z"]
-    }));
-    let v = ActionValidator::new(&meta);
-    let result = v.validate_input(&json!({"x": 1}));
-    assert!(!result.is_valid());
-    assert_eq!(result.errors.len(), 2); // missing y, z
-}
-
-// ── properties ────────────────────────────────────────────────────────────
-
-#[test]
-fn test_property_type_check() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "properties": {
-            "radius": { "type": "number" }
-        }
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!({"radius": 1.5})).is_valid());
-    let result = v.validate_input(&json!({"radius": "big"}));
-    assert!(!result.is_valid());
-    assert!(result.errors[0].path.contains("radius"));
-}
-
-// ── numeric constraints ───────────────────────────────────────────────────
-
-#[test]
-fn test_minimum_passes() {
-    let meta = make_meta_with_schema(json!({
-        "type": "number",
-        "minimum": 0.0
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(0.0)).is_valid());
-    assert!(v.validate_input(&json!(100.0)).is_valid());
-}
-
-#[test]
-fn test_minimum_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "number", "minimum": 0.0 }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!(-0.1)).is_valid());
-}
-
-#[test]
-fn test_maximum_passes() {
-    let meta = make_meta_with_schema(json!({ "type": "number", "maximum": 100.0 }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(50.0)).is_valid());
-    assert!(v.validate_input(&json!(100.0)).is_valid());
-}
-
-#[test]
-fn test_maximum_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "number", "maximum": 100.0 }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!(100.1)).is_valid());
-}
-
-#[test]
-fn test_min_max_range() {
-    let meta = make_meta_with_schema(json!({ "minimum": -1.0, "maximum": 1.0 }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(0.5)).is_valid());
-    assert!(!v.validate_input(&json!(2.0)).is_valid());
-    assert!(!v.validate_input(&json!(-2.0)).is_valid());
-}
-
-// ── string constraints ────────────────────────────────────────────────────
-
-#[test]
-fn test_max_length_passes() {
-    let meta = make_meta_with_schema(json!({ "type": "string", "maxLength": 5 }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!("hello")).is_valid()); // exactly 5
-    assert!(v.validate_input(&json!("hi")).is_valid());
-}
-
-#[test]
-fn test_max_length_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "string", "maxLength": 3 }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!("toolong")).is_valid());
-}
-
-#[test]
-fn test_min_length_passes() {
-    let meta = make_meta_with_schema(json!({ "type": "string", "minLength": 3 }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!("abc")).is_valid());
-}
-
-#[test]
-fn test_min_length_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "string", "minLength": 5 }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!("hi")).is_valid());
-}
-
-// ── array constraints ─────────────────────────────────────────────────────
-
-#[test]
-fn test_min_items_passes() {
-    let meta = make_meta_with_schema(json!({ "type": "array", "minItems": 2 }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!([1, 2])).is_valid());
-    assert!(v.validate_input(&json!([1, 2, 3])).is_valid());
-}
-
-#[test]
-fn test_min_items_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "array", "minItems": 2 }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!([1])).is_valid());
-    assert!(!v.validate_input(&json!([])).is_valid());
-}
-
-#[test]
-fn test_max_items_fails() {
-    let meta = make_meta_with_schema(json!({ "type": "array", "maxItems": 2 }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!([1, 2, 3])).is_valid());
-}
-
-#[test]
-fn test_items_type_check() {
-    let meta = make_meta_with_schema(json!({
-        "type": "array",
-        "items": { "type": "number" }
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!([1.0, 2.0])).is_valid());
-    let result = v.validate_input(&json!([1.0, "bad"]));
-    assert!(!result.is_valid());
-    assert!(result.errors[0].path.contains("[1]"));
-}
-
-// ── enum ──────────────────────────────────────────────────────────────────
-
-#[test]
-fn test_enum_passes() {
-    let meta = make_meta_with_schema(json!({ "enum": ["low", "medium", "high"] }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!("medium")).is_valid());
-}
-
-#[test]
-fn test_enum_fails() {
-    let meta = make_meta_with_schema(json!({ "enum": ["low", "medium", "high"] }));
-    let v = ActionValidator::new(&meta);
-    assert!(!v.validate_input(&json!("extreme")).is_valid());
-}
-
-// ── additionalProperties ──────────────────────────────────────────────────
-
-#[test]
-fn test_additional_properties_false_rejects_extra() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "properties": { "name": { "type": "string" } },
-        "additionalProperties": false
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!({"name": "x"})).is_valid());
-    let result = v.validate_input(&json!({"name": "x", "unknown": 1}));
-    assert!(!result.is_valid());
-    assert!(result.errors[0].message.contains("unknown"));
-}
-
-#[test]
-fn test_additional_properties_true_allows_extra() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "properties": { "name": { "type": "string" } },
-        "additionalProperties": true
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(
-        v.validate_input(&json!({"name": "x", "extra": 99}))
-            .is_valid()
-    );
-}
-
-// ── nested ────────────────────────────────────────────────────────────────
-
-#[test]
-fn test_nested_object_validation() {
-    let meta = make_meta_with_schema(json!({
-        "type": "object",
-        "required": ["position"],
-        "properties": {
-            "position": {
-                "type": "object",
-                "required": ["x", "y", "z"],
-                "properties": {
-                    "x": { "type": "number" },
-                    "y": { "type": "number" },
-                    "z": { "type": "number" }
-                }
-            }
-        }
-    }));
-    let v = ActionValidator::new(&meta);
-    assert!(
-        v.validate_input(&json!({"position": {"x": 1.0, "y": 2.0, "z": 3.0}}))
-            .is_valid()
-    );
-    let result = v.validate_input(&json!({"position": {"x": 1.0}}));
-    assert!(!result.is_valid());
-    // Missing y and z
-    assert!(result.errors.len() >= 2);
-}
-
-// ── empty / trivial schema ────────────────────────────────────────────────
-
-#[test]
-fn test_empty_schema_accepts_anything() {
-    let meta = make_meta_with_schema(json!({}));
-    let v = ActionValidator::new(&meta);
-    assert!(v.validate_input(&json!(null)).is_valid());
-    assert!(v.validate_input(&json!({"any": "thing"})).is_valid());
-}
-
-// ── ValidationResult helpers ──────────────────────────────────────────────
-
-#[test]
-fn test_validation_result_into_result_ok() {
-    let r = ValidationResult { errors: vec![] };
-    assert!(r.into_result().is_ok());
-}
-
-#[test]
-fn test_validation_result_into_result_err() {
-    let r = ValidationResult {
-        errors: vec![ValidationError {
-            path: "params.x".into(),
-            message: "bad".into(),
-        }],
-    };
-    let err = r.into_result().unwrap_err();
-    assert!(err.contains("params.x"));
-}
-
-// ── from_schema ───────────────────────────────────────────────────────────
-
-#[test]
-fn test_from_schema_direct() {
-    let schema = json!({ "type": "string", "minLength": 1 });
-    let v = ActionValidator::from_schema(schema);
-    assert!(v.validate_input(&json!("hello")).is_valid());
-    assert!(!v.validate_input(&json!("")).is_valid());
-}
-
-// ── boolean schema (JSON Schema `true`/`false`) ───────────────────────────
-
-#[test]
-fn test_boolean_schema_true_accepts_anything() {
-    // `true` schema: any value is valid
-    let v = ActionValidator::from_schema(json!(true));
-    assert!(v.validate_input(&json!(null)).is_valid());
-    assert!(v.validate_input(&json!(42)).is_valid());
-    assert!(v.validate_input(&json!("hello")).is_valid());
-    assert!(v.validate_input(&json!({"x": 1})).is_valid());
-}
-
-#[test]
-fn test_boolean_schema_false_accepts_anything() {
-    // Our validator skips validation for non-object schemas (incl. `false`)
-    // so no errors are generated — matches lenient "unknown schema → skip" policy.
-    let v = ActionValidator::from_schema(json!(false));
-    assert!(v.validate_input(&json!(null)).is_valid());
-}
-
-// ── additionalProperties: false without properties ────────────────────────
-
-#[test]
-fn test_additional_properties_false_no_properties_key() {
-    // `additionalProperties: false` without a `properties` key —
-    // no known properties means every key is "additional".
-    let v = ActionValidator::from_schema(json!({
-        "type": "object",
-        "additionalProperties": false
-    }));
-    // An empty object has no additional properties → valid
-    assert!(v.validate_input(&json!({})).is_valid());
-    // Any key is extra → invalid
-    let result = v.validate_input(&json!({"anything": 1}));
-    assert!(!result.is_valid());
-}
-
-// ── deeply nested array items ─────────────────────────────────────────────
-
-#[test]
-fn test_nested_array_items_type_check() {
-    // Array of arrays of numbers
-    let v = ActionValidator::from_schema(json!({
-        "type": "array",
-        "items": {
-            "type": "array",
-            "items": { "type": "number" }
-        }
-    }));
-    assert!(v.validate_input(&json!([[1.0, 2.0], [3.0]])).is_valid());
-    let result = v.validate_input(&json!([[1.0, "bad"], [3.0]]));
-    assert!(!result.is_valid());
-    // Path should include both array indices
-    assert!(result.errors[0].path.contains("[0]"));
-    assert!(result.errors[0].path.contains("[1]"));
-}
-
-// ── enum inside property ──────────────────────────────────────────────────
-
-#[test]
-fn test_enum_inside_property() {
-    let v = ActionValidator::from_schema(json!({
-        "type": "object",
-        "properties": {
-            "mode": { "enum": ["read", "write", "append"] }
-        }
-    }));
-    assert!(v.validate_input(&json!({"mode": "read"})).is_valid());
-    let result = v.validate_input(&json!({"mode": "invalid"}));
-    assert!(!result.is_valid());
-    assert!(result.errors[0].path.contains("mode"));
-}
-
-// ── multiple errors accumulate ────────────────────────────────────────────
-
-#[test]
-fn test_multiple_errors_accumulate() {
-    let v = ActionValidator::from_schema(json!({
-        "type": "object",
-        "required": ["x", "y"],
-        "properties": {
-            "x": { "type": "number", "minimum": 0.0 },
-            "y": { "type": "string", "maxLength": 5 }
-        }
-    }));
-    // x wrong type, y too long
-    let result = v.validate_input(&json!({"x": "bad", "y": "toolong"}));
-    assert!(!result.is_valid());
-    // Should have at least 2 errors (type mismatch for x + length for y)
-    assert!(result.errors.len() >= 2);
-}
-
-// ── array with both minItems and items schema ─────────────────────────────
-
-#[test]
-fn test_array_min_items_and_items_schema() {
-    let v = ActionValidator::from_schema(json!({
-        "type": "array",
-        "minItems": 2,
-        "items": { "type": "number" }
-    }));
-    assert!(v.validate_input(&json!([1.0, 2.0])).is_valid());
-    // Too short
-    assert!(!v.validate_input(&json!([1.0])).is_valid());
-    // Correct length but wrong item type
-    let result = v.validate_input(&json!([1.0, "oops"]));
-    assert!(!result.is_valid());
-}
-
-// ── string boundary at exact limits ──────────────────────────────────────
-
-#[test]
-fn test_string_exact_min_max_length() {
-    let v = ActionValidator::from_schema(json!({
-        "type": "string",
-        "minLength": 3,
-        "maxLength": 5
-    }));
-    assert!(v.validate_input(&json!("abc")).is_valid()); // exactly 3
-    assert!(v.validate_input(&json!("abcde")).is_valid()); // exactly 5
-    assert!(!v.validate_input(&json!("ab")).is_valid()); // 2 < min
-    assert!(!v.validate_input(&json!("abcdef")).is_valid()); // 6 > max
-}
-
-// ── numeric boundary at exact limits ─────────────────────────────────────
-
-#[test]
-fn test_numeric_exact_min_max() {
-    let v = ActionValidator::from_schema(json!({
-        "type": "number",
-        "minimum": 0.0,
-        "maximum": 1.0
-    }));
-    assert!(v.validate_input(&json!(0.0)).is_valid()); // exactly min
-    assert!(v.validate_input(&json!(1.0)).is_valid()); // exactly max
-    assert!(!v.validate_input(&json!(-0.001)).is_valid());
-    assert!(!v.validate_input(&json!(1.001)).is_valid());
-}
-
-// ── error message format ──────────────────────────────────────────────────
-
-#[test]
-fn test_validation_error_display() {
-    let e = ValidationError {
-        path: "params.radius".into(),
-        message: "must be >= 0".into(),
-    };
-    assert_eq!(e.to_string(), "params.radius: must be >= 0");
-}
