+++ crates/dcc-mcp-actions/src/validator/tests/mod.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+mod advanced;
+mod constraints;
+pub(super) mod fixtures;
+mod object_schema;
+mod type_checks;
